requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
biopython>=1.79
torch>=1.9.0
transformers>=4.10.0
scipy>=1.7.0
joblib>=1.0.0
tqdm>=4.62.0
